#227 fix problem at update users without valid media folders (#226)
UserSettingsController.updateUser(...) receives a UserSettingsCommand
with allowedMusicFolderIds set to NULL whenever neither enabled nor
existing folders are present in the system. The form in
userSettings.jsp will not display any selection and therefore, will
also leave that field unpopulated in the POST request.

This lead to user management being broken for unconfigured or
misconfigured systems.